+++ test/create-instance.js
@@ -11,10 +11,10 @@ require('./common');
 
 let User = nano_orm.defineModel('user', ['username', 'password']);
 
-describe('create', () => {
+describe('constructor', () => {
 
 	it('No params results in default values for all parameters', () => {
-		let user = User.create();
+		let user = new User();
 
 		expect(user.id      ).is.deep.equal(0);
 		expect(user.username).is.deep.equal(null);
@@ -22,7 +22,7 @@ describe('create', () => {
 	});
 
 	it('Incomplete params results in default values for only missing parameters', () => {
-		let user = User.create({password: '999'});
+		let user = new User({password: '999'});
 
 		expect(user.id      ).is.deep.equal(0);
 		expect(user.username).is.deep.equal(null);
@@ -31,7 +31,7 @@ describe('create', () => {
 
 	it('Complete params fills in all data fields of instance', () => {
 
-		let user = User.create({password: '123', username: 'Jim'});
+		let user = new User({password: '123', username: 'Jim'});
 
 		expect(user.id      ).is.deep.equal(0);
 		expect(user.username).is.deep.equal('Jim');
@@ -74,5 +74,45 @@ describe('createFromRow', () => {
 
 	it('Row with missing id field results in exception', () => {
 		expect(() => User.createFromRow({ password: '', username: '' })).to.throw();
+	});
+});
+
+describe('createFromRows', () => {
+	it('Array of length 0 results in empty array', () => {
+		let user = User.createFromRows([]);
+		expect(user).is.deep.equal([]);
+	});
+
+	it('Array of length 1 results in single valid instance', () => {
+		let user = User.createFromRows([
+			{ id: 10, username: 'Tim', password: 'stuff' }
+		]);
+
+		expect(user[0].id      ).is.deep.equal(10     );
+		expect(user[0].username).is.deep.equal('Tim'  );
+		expect(user[0].password).is.deep.equal('stuff');
+	});
+
+	it('Array of length 2 results in 2 valid instances', () => {
+		let user = User.createFromRows([
+			{ id: 10, username: 'Tim', password: 'stuff' },
+			{ id: 11, username: 'Bob', password: 'thing' }
+		]);
+
+		expect(user[0].id      ).is.deep.equal(10     );
+		expect(user[0].username).is.deep.equal('Tim'  );
+		expect(user[0].password).is.deep.equal('stuff');
+
+		expect(user[1].id      ).is.deep.equal(11     );
+		expect(user[1].username).is.deep.equal('Bob'  );
+		expect(user[1].password).is.deep.equal('thing');
+	});
+
+	it('Array containing invalid objects throws error', () => {
+		expect(() => User.createFromRows([
+			{ id: 10, username: 'Tim', password: 'stuff' },
+			{ id: 11, username: 'Bob', password: 'thing' },
+			{},
+		])).to.throw();
 	});
 });